+++ lib/lastfm_archive/livebook.ex
@@ -2,13 +2,21 @@ defmodule LastfmArchive.Livebook do
   @moduledoc """
   Livebook chart and text rendering.
   """
-
   alias LastfmArchive.LastfmClient.Impl, as: LastfmClient
   alias LastfmArchive.LastfmClient.LastfmApi
   alias VegaLite, as: Vl
 
+  alias Explorer.DataFrame
+  alias Explorer.Series
+
+  require Explorer.DataFrame
+
   @cache LastfmArchive.Cache.Server
-  @type monthy_count :: %{count: integer(), date: String.t()}
+  @type user :: LastfmArchive.Behaviour.Archive.user()
+  @type year :: integer()
+  @type daily_playcounts :: %{
+          {user, year} => %{data: list(%{count: integer(), date: String.t()}), total: integer(), max: integer()}
+        }
 
   @doc """
   Display user name and total number of scrobbles to archive.
@@ -41,55 +49,99 @@ defmodule LastfmArchive.Livebook do
   end
 
   @doc """
-  Monthly playcounts of scrobbles archived so far - VegaLite heatmap.
+  Display daily playcounts of scrobbles archived in VegaLite heatmaps.
   """
-  @spec monthly_playcounts_heatmap(module()) :: VegaLite.t()
-  def monthly_playcounts_heatmap(cache \\ @cache) do
-    Vl.new(title: "")
-    |> Vl.data_from_values(status(cache))
-    |> Vl.mark(:rect)
+  @spec render_playcounts_heatmaps(user(), keyword(), module()) :: :ok
+  def render_playcounts_heatmaps(user \\ LastfmClient.default_user(), opts \\ [], cache \\ @cache) do
+    colour_scheme = Keyword.get(opts, :colour, "yellowgreenblue")
+    stats = daily_playcounts_per_years(user, cache)
+    global_max = stats |> Map.values() |> Stream.map(& &1.max) |> Enum.max()
+
+    stats
+    |> Enum.each(fn {{_user, year}, %{data: data}} ->
+      render_heading(year, data)
+      render_heatmap(data, year, global_max, colour_scheme)
+    end)
+  end
+
+  defp render_heading(year, data) do
+    stats = stats_per_year(data)
+
+    Kino.Markdown.new("<small style=\"padding-left: 40px;\">Year <b>#{year}</b>,
+      total <b>#{stats["total"]}</b>,
+      per-day
+      <b>#{stats["avg"] |> round()}</b> avg,
+      <b>#{stats["median"] |> round()}</b> median,
+      <b>#{stats["min"]}</b> min,
+      <b>#{stats["max"]}</b> max
+      </small>")
+    |> Kino.render()
+  end
+
+  defp stats_per_year(counts) do
+    DataFrame.new(counts)
+    |> DataFrame.collect()
+    |> DataFrame.summarise(
+      avg: mean(count),
+      median: median(count),
+      min: min(count),
+      max: max(count),
+      total: sum(count)
+    )
+    |> DataFrame.to_rows()
+    |> hd
+  end
+
+  defp render_heatmap(data, year, max, colour_scheme) do
+    Vl.new(title: nil, width: 620, height: 80)
+    |> Vl.transform(filter: "datum.count > 0 && datum.year == #{year}")
+    |> Vl.data_from_values(data)
+    |> Vl.mark(:rect, width: 9, height: 9, tooltip: true)
     |> Vl.encode_field(:x, "date",
-      time_unit: :month,
-      type: :ordinal,
-      title: "Month",
-      axis: [label_angle: 0, format: "%m"]
+      time_unit: :yearweek,
+      type: :temporal,
+      title: nil,
+      axis: [format: "%b", offset: -8, domain: false, grid: false],
+      scale: [domain: [[year: year, month: "jan", date: 1], [year: year, month: "dec", date: 31]]]
     )
     |> Vl.encode_field(:y, "date",
-      time_unit: :year,
-      type: :ordinal,
-      title: "Year"
+      time_unit: :yearday,
+      type: :temporal,
+      title: nil,
+      axis: [format: "%a", offset: 16],
+      sort: "descending"
     )
     |> Vl.encode_field(:color, "count",
-      aggregate: :max,
+      aggregate: :sum,
       type: :quantitative,
-      legend: [title: nil]
+      legend: false,
+      scale: [domain: [0, max], scheme: colour_scheme]
     )
+    |> Vl.encode(:tooltip, [[field: "date", type: :temporal], [field: "count", type: :quantitative]])
     |> Vl.config(view: [stroke: nil])
+    |> Kino.render()
   end
 
   @doc """
-  Yearly playcounts of scrobbles archived so far - Kino data table.
+  Returns a list of daily scrobble playcounts and stats per years.
   """
-  @spec yearly_playcounts_table(module()) :: Kino.JS.Live.t()
-  def yearly_playcounts_table(cache \\ @cache) do
-    status(cache, granularity: :yearly)
-    |> Kino.DataTable.new(keys: [:year, :count], name: "")
+  @spec daily_playcounts_per_years(user(), module()) :: daily_playcounts()
+  def daily_playcounts_per_years(user \\ LastfmClient.default_user(), cache \\ @cache) do
+    for {{user, year}, statuses} <- user |> LastfmArchive.Cache.load(cache), into: %{} do
+      data = aggregate(statuses) |> List.flatten()
+
+      {
+        {user, year},
+        %{
+          data: data,
+          max: Stream.map(data, & &1.count) |> Enum.max(),
+          total: Stream.map(data, & &1.count) |> Enum.sum()
+        }
+      }
+    end
   end
 
-  @doc """
-  Returns a list of monthly playcounts of scrobbles archived so far.
-  """
-  @spec status(module(), keyword()) :: list(monthy_count)
-  def status(cache \\ @cache, options \\ []) do
-    granularity = Keyword.get(options, :granularity, :monthly)
-
-    LastfmClient.default_user()
-    |> LastfmArchive.Cache.load(cache)
-    |> Enum.map(fn {{_user, year}, statuses} -> aggregate_counts(year, statuses, granularity) end)
-    |> List.flatten()
-  end
-
-  defp aggregate_counts(_year, statuses, :monthly) do
+  defp aggregate(statuses) do
     statuses
     |> Enum.flat_map(fn {{from, _to}, {count, status}} ->
       case Enum.all?(status, &(&1 == :ok)) do
@@ -97,24 +149,74 @@ defmodule LastfmArchive.Livebook do
         false -> []
       end
     end)
-    |> Enum.group_by(fn {from, _count} -> first_day_of_month(from) end, &elem(&1, 1))
-    |> Enum.into([], fn {day, counts} -> %{date: day, count: Enum.sum(counts)} end)
-  end
-
-  defp aggregate_counts(year, statuses, :yearly) do
-    statuses
-    |> Enum.flat_map(fn {{_from, _to}, {count, status}} ->
-      case Enum.all?(status, &(&1 == :ok)) do
-        true -> [count]
-        false -> []
-      end
+    |> Enum.group_by(fn {from, _count} -> get_date(from) end, &elem(&1, 1))
+    |> Enum.into([], fn {%{year: year} = date, counts} ->
+      %{
+        date: date |> to_string(),
+        year: year,
+        count: Enum.sum(counts)
+      }
     end)
-    |> Enum.sum()
-    |> then(fn playcount -> %{year: year, count: playcount} end)
   end
 
-  defp first_day_of_month(datetime) do
-    %DateTime{month: month, year: year} = datetime |> DateTime.from_unix!()
-    %Date{month: month, year: year, day: 1} |> to_string()
+  defp get_date(datetime), do: datetime |> DateTime.from_unix!() |> DateTime.to_date()
+
+  @doc """
+  Display faceted dataframe in VegaLite bubble plot showing first play and counts (size, colour).
+  """
+  @spec render_first_play_bubble_plot(DataFrame.t()) :: VegaLite.t()
+  def render_first_play_bubble_plot(facet_dataframe) do
+    {min_year, max_year} = find_first_play_min_max_years(facet_dataframe)
+
+    data =
+      facet_dataframe
+      |> DataFrame.put(:first_play, Series.cast(facet_dataframe[:first_play], :date) |> Series.cast(:string))
+      |> DataFrame.to_rows()
+
+    Vl.new(title: nil, width: 800, height: 400)
+    |> Vl.transform(calculate: "random()", as: "jitter")
+    |> Vl.transform(filter: "datum.counts > 0")
+    |> Vl.data_from_values(data)
+    |> Vl.mark(:circle, tooltip: true)
+    |> Vl.encode_field(:x, "first_play",
+      time_unit: :yearmonth,
+      type: :temporal,
+      title: nil,
+      axis: [format: "%Y"],
+      scale: [domain: [[year: min_year, month: "jan", date: 1], [year: max_year, month: "dec", date: 31]]]
+    )
+    |> Vl.encode_field(:x_offset, "jitter", type: :quantitative)
+    |> Vl.encode_field(:y_offset, "jitter", type: :quantitative)
+    |> Vl.encode_field(:y, "first_play",
+      time_unit: :date,
+      type: :temporal,
+      title: nil,
+      axis: [format: "%d"]
+      # sort: "ascending"
+    )
+    |> Vl.encode_field(:size, "counts",
+      type: :quantitative,
+      scale: [type: "linear", range_max: 1000, range_min: 2],
+      legend: false
+    )
+    |> Vl.encode_field(:color, "counts",
+      type: :nominal,
+      opacity: 0.5,
+      scale: [scheme: "turbo"],
+      legend: false
+    )
+    |> Vl.encode(:tooltip, [
+      [field: "artist", type: :nominal],
+      [field: "counts", type: :quantitative],
+      [field: "first_play", type: :temporal]
+    ])
+  end
+
+  defp find_first_play_min_max_years(df) do
+    df = df |> DataFrame.summarise(min: min(first_play), max: max(first_play)) |> DataFrame.to_rows() |> hd
+    %{year: min_year} = df["min"]
+    %{year: max_year} = df["max"]
+
+    {min_year, max_year}
   end
 end