+++ test/lastfm_archive/archive/transformers/transformer_test.exs
@@ -6,13 +6,15 @@ defmodule LastfmArchive.Archive.Transformers.TransformerTest do
 
   alias LastfmArchive.Archive.FileArchiveMock
   alias LastfmArchive.Archive.Transformers.Transformer
+  alias LastfmArchive.Archive.Transformers.TransformerConfigs
   alias LastfmArchive.FileIOMock
 
   import ExUnit.CaptureLog
   import Hammox
 
+  import LastfmArchive.Archive.Transformers.Transformer, only: [validate_opts: 1]
   import LastfmArchive.Factory, only: [build: 2, scrobbles_csv_gzipped: 0, dataframe: 0]
-  import LastfmArchive.Utils, only: [user_dir: 1]
+  import LastfmArchive.Utils.Archive, only: [derived_archive_dir: 1, user_dir: 1]
 
   require Explorer.DataFrame
 
@@ -130,7 +132,7 @@ defmodule LastfmArchive.Archive.Transformers.TransformerTest do
       format = :csv
       options = [format: format]
       write_opts = Transformer.write_opts(format)
-      archive_dir = "#{Transformer.derived_archive_dir(options)}"
+      archive_dir = "#{options |> validate_opts() |> derived_archive_dir()}"
 
       filepath = Path.join([user_dir(user), archive_dir, "2022.#{format}.gz"])
 
@@ -172,7 +174,7 @@ defmodule LastfmArchive.Archive.Transformers.TransformerTest do
           format = unquote(format)
           options = [format: format]
           write_opts = Transformer.write_opts(format)
-          archive_dir = "#{Transformer.derived_archive_dir(options)}"
+          archive_dir = "#{options |> validate_opts() |> derived_archive_dir()}"
 
           filepath1 = Path.join([user_dir(user), archive_dir, "2022.#{format}"])
           filepath2 = Path.join([user_dir(user), archive_dir, "2023.#{format}"])
@@ -256,7 +258,7 @@ defmodule LastfmArchive.Archive.Transformers.TransformerTest do
   describe "apply/3" do
     test "transform all years", %{dataframe: df, metadata: metadata, transformer: transformer} do
       options = [format: :ipc_stream]
-      archive_dir = "#{Transformer.derived_archive_dir(options)}"
+      archive_dir = "#{options |> validate_opts() |> derived_archive_dir()}"
       dir = Path.join(user_dir(metadata.creator), archive_dir)
 
       FileIOMock
@@ -276,12 +278,13 @@ defmodule LastfmArchive.Archive.Transformers.TransformerTest do
 
     test "transform a given year", %{dataframe: df, metadata: metadata, transformer: transformer} do
       options = [format: :ipc_stream, year: 2022]
-      archive_dir = "#{Transformer.derived_archive_dir(options)}"
+      archive_dir = "#{options |> validate_opts() |> derived_archive_dir()}"
       dir = Path.join(user_dir(metadata.creator), archive_dir)
+      filepath = "#{dir}/2022.ipc_stream"
 
       FileIOMock
       |> expect(:exists?, fn ^dir -> true end)
-      |> expect(:exists?, fn _filepath -> false end)
+      |> expect(:exists?, fn ^filepath -> false end)
 
       # source 1-year, 12 months scrobbles data
       FileArchiveMock |> expect(:read, 12, fn ^metadata, _options -> {:ok, df} end)
@@ -291,6 +294,34 @@ defmodule LastfmArchive.Archive.Transformers.TransformerTest do
       capture_log(fn ->
         assert {:ok, %LastfmArchive.Archive.Metadata{}} = Transformer.apply(transformer, metadata, options)
       end)
+    end
+  end
+
+  describe "validate_opts/1" do
+    test "returns defaults options" do
+      opts = []
+
+      for opt <- Transformer.validate_opts(opts) do
+        assert opt in TransformerConfigs.default_opts()
+      end
+    end
+
+    test "use option(s) if given" do
+      opts = [format: :csv]
+
+      assert Transformer.validate_opts(opts) |> length() == TransformerConfigs.default_opts() |> length()
+
+      for opt <- Transformer.validate_opts(opts) do
+        assert opt in (TransformerConfigs.default_opts() |> Keyword.replace!(:format, :csv))
+      end
+    end
+
+    test "remove unrelated options" do
+      opts = [format: :csv, non_transformer_opt: 1234]
+
+      assert Transformer.validate_opts(opts) |> length() == TransformerConfigs.default_opts() |> length()
+      assert {:format, :csv} in Transformer.validate_opts(opts)
+      refute {:non_transformer_opt, 1234} in Transformer.validate_opts(opts)
     end
   end
 end