+++ test/lastfm_archive/livebook_test.exs
@@ -7,7 +7,9 @@ defmodule LastfmArchive.LivebookTest do
   alias LastfmArchive.Cache
   alias LastfmArchive.Cache.Server, as: CacheServer
   alias LastfmArchive.Livebook, as: LFM_LB
-  alias LastfmArchive.Utils
+
+  import LastfmArchive.Factory, only: [dataframe: 1]
+  import LastfmArchive.Utils.Archive, only: [user_dir: 1]
 
   @cache :livebooktest_cache
 
@@ -18,7 +20,7 @@ defmodule LastfmArchive.LivebookTest do
     {:ok, _pid} = start_supervised(%{id: @cache, start: {CacheServer, :start_link, [[name: @cache]]}})
     cache = %{{"a_user", 2006} => %{{1_138_752_000, 1_138_838_399} => {33, [:ok]}}}
 
-    cache_file = "#{Utils.user_dir("a_user")}/#{Cache.cache_dir()}/2006"
+    cache_file = "#{user_dir("a_user")}/#{Cache.cache_dir()}/2006"
     stub_with(LastfmClient.impl(), LastfmArchive.LastfmClientStub)
     LastfmArchive.PathIOMock |> stub(:wildcard, fn _, _ -> [cache_file] end)
     LastfmArchive.FileIOMock |> stub(:read!, fn _ -> cache[{"a_user", 2006}] |> :erlang.term_to_binary() end)
@@ -30,15 +32,11 @@ defmodule LastfmArchive.LivebookTest do
     assert %Kino.Markdown{} = LFM_LB.info()
   end
 
-  test "monthly_playcounts_heatmap/1" do
-    assert %VegaLite{} = LFM_LB.monthly_playcounts_heatmap(@cache)
+  test "render_playcounts_heatmaps/1" do
+    assert :ok = LFM_LB.render_playcounts_heatmaps("a_lastfm_user", [], @cache)
   end
 
-  test "yearly_playcounts_table/1" do
-    assert %Kino.JS.Live{module: Kino.Table} = LFM_LB.yearly_playcounts_table(@cache)
-  end
-
-  test "status/0" do
-    assert [%{count: 33, date: "2006-02-01"}] == LFM_LB.status(@cache)
+  test "render_first_play_bubble_plot/1" do
+    assert %VegaLite{} = LFM_LB.render_first_play_bubble_plot(dataframe(:artists) |> Explorer.DataFrame.collect())
   end
 end