+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # Changelog
+## 1.2.0 (2023-09-27)
+
+* `FacetsTransformer` and `facet` (unique artists, albums, tracks) option in `transform/2` for facets archiving
+* refactor: utils modules, dataframe and transformer config macros
+* Livebook: 
+    * [`Creating a file archive`](https://hexdocs.pm/lastfm_archive/archiving.html): heatmaps visualisation containing daily playcounts, tooltips and stats
+    * [`Facets archiving`](https://hexdocs.pm/lastfm_archive/facets.html) new guide about facets archiving with usage demo for `Explorer.DataFrame` analytics and [Vega-Lite](https://vega.github.io/vega-lite/) visualisation
+
 ## 1.1.1 (2023-09-15)
 
 * fix: Lastfm client bug that fetches page 1 instead of page 2 (for scrobbles > 200 per-day/on the day)